Reduce code repetition and file numbers
